move get_tokens to a separate file
