+++ find_duplicates_test.py
@@ -3,14 +3,14 @@ import numpy
 import unittest
 
 import find_duplicates
+import tokenizer
 import utils
-import visual_diff
 
 
 class TestGetLengths(unittest.TestCase):
     def setUp(self):
         contents = 'print(hello(1, 2)) and print(hello("hi", 2))'
-        data = visual_diff.get_tokens(contents, "python")
+        data = tokenizer.get_tokens(contents, "python")
         self.matrix = utils.make_matrix(data.tokens, data.tokens)
 
     @staticmethod
